requirements.txt
openai
google-generativeai
huggingface_hub
langchain_huggingface
python-dotenv
streamlit